:recycle: Refactoring: Building class 구조 변경 (#6)
