+++ src/main/java/forus/naviforyou/global/common/collection/building/Building.java
@@ -1,5 +1,6 @@
-package forus.naviforyou.global.common.entity;
+package forus.naviforyou.global.common.collection.building;
 
+import forus.naviforyou.global.common.collection.enums.Accessibility;
 import lombok.*;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.DBRef;
@@ -19,15 +20,19 @@ public class Building {
     private String id;
 
     private String name;
-    private String generalAddress;
-    private String roadAddress;
-    private String detailAddress;
-    private Float posX;
-    private Float posY;
 
-    @DBRef
-    @Builder.Default
-    private List<CustomizedAccessibility> customizedAccessibilityList = new ArrayList<>();
+    private String generalAddress;
+
+    private String roadAddress;
+
+    private String detailAddress;
+
+    private Location location ;
+
+    private List<Accessibility> accessibilityList ; // 보유한 장애물 리스트
+
+    private List<UserReport> userReportList ; // 유저가 추가한 {장애물,count} 리스트
+
 
 }
