+++ src/main/java/forus/naviforyou/global/common/collection/building/Location.java
@@ -1,0 +1,6 @@
+package forus.naviforyou.global.common.collection.building;
+
+public class Location {
+    private Float posX;
+    private Float posY;
+}
