+++ src/main/java/forus/naviforyou/global/common/collection/building/UserReport.java
@@ -1,0 +1,8 @@
+package forus.naviforyou.global.common.collection.building;
+
+import forus.naviforyou.global.common.collection.enums.Accessibility;
+
+public class UserReport {
+    private Accessibility accessibility;
+    private int num;
+}
